# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Uganda Diaspora Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news articles
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string } }
        - { name: featured, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: News list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsListResponse"
    post:
      operationId: createNews
      tags: [news]
      summary: Create news article
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
  /news/{id}:
    get:
      operationId: getNews
      tags: [news]
      summary: Get news article
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
    patch:
      operationId: updateNews
      tags: [news]
      summary: Update news article
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsUpdate"
      responses:
        "200":
          description: Updated article
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete news article
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List community posts
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: search, in: query, schema: { type: string } }
      responses:
        "200":
          description: Posts list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{id}:
    get:
      operationId: getPost
      tags: [posts]
      summary: Get post
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "404":
          description: Not found
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete post
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Toggle like on post
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Post with updated likes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"

  # COMMENTS
  /posts/{postId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for post
      parameters:
        - { name: postId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Create comment
      parameters:
        - { name: postId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /posts/{postId}/comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete comment
      parameters:
        - { name: postId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # NOTIFICATIONS
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
    post:
      operationId: sendNotification
      tags: [notifications]
      summary: Send push notification (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete notification
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # MDAs
...
  /opportunities:
    get:
      operationId: listOpportunities
      tags: [opportunities]
      summary: List opportunities
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: search, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { type: string } }
      responses:
        "200":
          description: Opportunities list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpportunityListResponse"
    post:
      operationId: createOpportunity
      tags: [opportunities]
      summary: Create opportunity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
  /opportunities/{id}:
    get:
      operationId: getOpportunity
      tags: [opportunities]
      summary: Get opportunity
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
        "404":
          description: Not found
    patch:
      operationId: updateOpportunity
      tags: [opportunities]
      summary: Update opportunity
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpportunityUpdate"
      responses:
        "200":
          description: Updated opportunity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Opportunity"
    delete:
      operationId: deleteOpportunity
      tags: [opportunities]
      summary: Delete opportunity
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # DASHBOARD
...
components:
  schemas:
...
    NewsArticle:
      type: object
      required: [id, title, content, category, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        content: { type: string }
        summary: { type: ["string", "null"] }
        category: { type: string }
        imageUrl: { type: ["string", "null"] }
        isFeatured: { type: boolean }
        isPublished: { type: boolean }
        authorName: { type: ["string", "null"] }
        viewCount: { type: integer }
        createdAt: { type: string, format: date-time }
        publishedAt: { type: ["string", "null"] }
    NewsInput:
      type: object
      required: [title, content, category]
      properties:
        title: { type: string }
        content: { type: string }
        summary: { type: string }
        category: { type: string }
        imageUrl: { type: string }
        isFeatured: { type: boolean }
        isPublished: { type: boolean }
        authorName: { type: string }
    NewsUpdate:
      type: object
      properties:
        title: { type: string }
        content: { type: string }
        summary: { type: string }
        category: { type: string }
        imageUrl: { type: string }
        isFeatured: { type: boolean }
        isPublished: { type: boolean }
    NewsListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/NewsArticle"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
...
    Post:
      type: object
      required: [id, content, authorName, createdAt]
      properties:
        id: { type: integer }
        content: { type: string }
        imageUrl: { type: ["string", "null"] }
        authorName: { type: string }
        authorAvatarUrl: { type: ["string", "null"] }
        likeCount: { type: integer }
        commentCount: { type: integer }
        isModerated: { type: boolean }
        createdAt: { type: string, format: date-time }
    PostInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
        imageUrl: { type: string }
    PostListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # COMMENTS
    Comment:
      type: object
      required: [id, content, authorName, createdAt]
      properties:
        id: { type: integer }
        content: { type: string }
        authorName: { type: string }
        authorAvatarUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    CommentInput:
      type: object
      required: [content]
      properties:
        content: { type: string }

    # NOTIFICATIONS
    Notification:
      type: object
      required: [id, title, body, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        body: { type: string }
        type: { type: ["string", "null"] }
        audience: { type: ["string", "null"] }
        sentCount: { type: integer }
        createdAt: { type: string, format: date-time }
    NotificationInput:
      type: object
      required: [title, body]
      properties:
        title: { type: string }
        body: { type: string }
        type: { type: string }
        audience: { type: string }
    NotificationListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # MDAs
...
    Opportunity:
      type: object
      required: [id, title, type, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        type: { type: string }
        organization: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        deadline: { type: ["string", "null"] }
        applicationUrl: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    OpportunityInput:
      type: object
      required: [title, type]
      properties:
        title: { type: string }
        description: { type: string }
        type: { type: string }
        organization: { type: string }
        location: { type: string }
        deadline: { type: string }
        applicationUrl: { type: string }
        imageUrl: { type: string }
    OpportunityUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        type: { type: string }
        organization: { type: string }
        location: { type: string }
        deadline: { type: string }
        applicationUrl: { type: string }
        isActive: { type: boolean }
    OpportunityListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Opportunity"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # DASHBOARD
...