# lib/api-zod/src/generated/types/postListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Uganda Diaspora Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Post } from './post';

export interface PostListResponse {
  data: Post[];
  total: number;
  page: number;
  limit: number;
}